Move findReferences to utils. It will be used in extendedValidate too.
